+++ Common/Models/PortfolioModels/Properties/UpdateProperty.cs
@@ -1,20 +1,20 @@
 ﻿namespace Common.Models.PortfolioModels.Properties
 {
-    public class UpdateProperty
-    {
-        /// <summary>
-        /// Gets or sets the Id of the property.
-        /// </summary>
+	public class UpdateProperty
+	{
+		/// <summary>
+		/// Gets or sets the Id of the property.
+		/// </summary>
 		public int Id { get; set; }
 
-        /// <summary>
-        /// Gets or sets the name of the property.
-        /// </summary>
+		/// <summary>
+		/// Gets or sets the name of the property.
+		/// </summary>
 		public string Name { get; set; }
 
-        /// <summary>
-        /// Gets or sets the value of the property.
-        /// </summary>
-        public decimal Value { get; set; }
-    }
+		/// <summary>
+		/// Gets or sets the value of the property.
+		/// </summary>
+		public double Value { get; set; }
+	}
 }
